sockets: separation of concerns
