+++ apps/api/src/controllers/game/index.ts
@@ -1,0 +1,2 @@
+export * from './onJoin'
+export * from './onEndPhase'
