+++ apps/api/src/lib/roomGameData.ts
@@ -1,0 +1,9 @@
+import { RoomData } from '../types'
+
+/**
+ * roomGameData
+ *
+ * The singleton used to store the roomGameData
+ * used in controllers throughout the socket implementation.
+ * */
+export const roomGameData = new Map<string, RoomData>()
