+++ apps/api/src/modules/gameModule.ts
@@ -1,8 +1,6 @@
 import { Server } from 'socket.io'
-import {
-  onJoin as onJoinController,
-  onEndPhase as onEndPhaseController,
-} from '../controllers/gameController'
+import { onJoin } from '../controllers/game'
+import { onEndPhase } from '../controllers/game'
 
 const gameModule = (io: Server) => {
   const gameNamespace = io.of('/game')
@@ -10,11 +8,9 @@ const gameModule = (io: Server) => {
     console.info(`Player connected: ${socket.id}`)
 
     socket.on('game:join', payload =>
-      onJoinController(socket, payload, gameNamespace),
+      onJoin(socket, payload, gameNamespace),
     )
-    socket.on('game:endPhase', payload =>
-      onEndPhaseController(socket, payload),
-    )
+    socket.on('game:endPhase', payload => onEndPhase(socket, payload))
 
     socket.on('disconnect', () => {
       console.info(`Player disconnected: ${socket.id}`)
